The download works properly.
